+++ common/src/main/java/io/github/mortuusars/exposure/block/entity/Lightroom.java
@@ -10,5 +10,6 @@ public class Lightroom {
     public static final int BLACK_SLOT = 5;
     public static final int RESULT_SLOT = 6;
 
+    public static final int[] ALL_SLOTS = new int[]{ 0, 1, 2, 3, 4, 5, 6};
     public static final int[] OUTPUT_SLOTS = new int[]{6};
 }