+++ forge/src/main/java/io/github/mortuusars/exposure/forge/mixin/LightroomBlockEntityForgeMixin.java
@@ -1,0 +1,42 @@
+package io.github.mortuusars.exposure.forge.mixin;
+
+import io.github.mortuusars.exposure.block.entity.LightroomBlockEntity;
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction;
+import net.minecraft.world.WorldlyContainer;
+import net.minecraft.world.level.block.entity.BaseContainerBlockEntity;
+import net.minecraft.world.level.block.entity.BlockEntityType;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraftforge.common.capabilities.Capability;
+import net.minecraftforge.common.capabilities.ForgeCapabilities;
+import net.minecraftforge.common.util.LazyOptional;
+import net.minecraftforge.items.IItemHandler;
+import net.minecraftforge.items.wrapper.SidedInvWrapper;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Unique;
+
+@Mixin(value = LightroomBlockEntity.class, remap = false)
+public abstract class LightroomBlockEntityForgeMixin extends BaseContainerBlockEntity implements WorldlyContainer {
+    protected LightroomBlockEntityForgeMixin(BlockEntityType<?> type, BlockPos pos, BlockState blockState) {
+        super(type, pos, blockState);
+    }
+
+    @Unique
+    private final LazyOptional<? extends IItemHandler>[] exposure$handlers = SidedInvWrapper.create(this, Direction.values());
+
+    @Override
+    public <T> @NotNull LazyOptional<T> getCapability(@NotNull Capability<T> capability, @Nullable Direction facing) {
+        if (!this.remove && facing != null && capability == ForgeCapabilities.ITEM_HANDLER)
+            return exposure$handlers[facing.ordinal()].cast();
+        return super.getCapability(capability, facing);
+    }
+
+    @Override
+    public void setRemoved() {
+        super.setRemoved();
+        for (LazyOptional<? extends IItemHandler> handler : exposure$handlers)
+            handler.invalidate();
+    }
+}